+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/jvm/features/HubdleKotlinJvmFeaturesExtension.kt
@@ -9,12 +9,11 @@ import com.javiersc.hubdle.extensions.kotlin.features.shared.HubdleKotlinExtende
 import com.javiersc.hubdle.extensions.kotlin.features.shared.HubdleKotlinKotestDelegateFeatureExtension
 import com.javiersc.hubdle.extensions.kotlin.features.shared.HubdleKotlinSerializationDelegateFeatureExtension
 import com.javiersc.hubdle.extensions.kotlin.jvm.hubdleKotlinJvm
+import com.javiersc.hubdle.extensions.shared.features.HubdleGradleDelegateFeatureExtension
 import com.javiersc.hubdle.extensions.shared.features.HubdleIntellijDelegateFeatureExtension
 import com.javiersc.hubdle.extensions.shared.features.HubdleJavaApplicationDelegateFeatureExtension
 import com.javiersc.hubdle.extensions.shared.features.HubdleJavaVersionDelegateFeatureExtension
 import com.javiersc.hubdle.extensions.shared.features.HubdleJvmToolchainDelegateFeatureExtension
-import com.javiersc.hubdle.extensions.shared.features.gradle.HubdleGradlePluginDelegateFeatureExtension
-import com.javiersc.hubdle.extensions.shared.features.gradle.HubdleGradleVersionCatalogDelegateFeatureExtension
 import javax.inject.Inject
 import org.gradle.api.Project
 import org.gradle.api.provider.Property
@@ -36,8 +35,7 @@ constructor(
 }
 
 public interface HubdleKotlinJvmDelegateFeaturesExtension :
-    HubdleGradlePluginDelegateFeatureExtension,
-    HubdleGradleVersionCatalogDelegateFeatureExtension,
+    HubdleGradleDelegateFeatureExtension,
     HubdleIntellijDelegateFeatureExtension,
     HubdleJavaApplicationDelegateFeatureExtension,
     HubdleKotlinComposeDelegateFeatureExtension,